more isolation

+++ Makefile
@@ -63,7 +63,7 @@ install: create-dirs install-core install-session install-extra
 install-terminal:
 	[ -f terminal/d-term ] && install terminal/d-term $(DESTDIR)/usr/lib/sulin/dsl/d-term || true
 	install terminal/d-term.desktop $(DESTDIR)/usr/lib/sulin/dsl/
-	install terminal/d-term.py $(DESTDIR)/usr/lib/sulin/dsl/d-term
+	[ -f terminal/d-term ] || install terminal/d-term.py $(DESTDIR)/usr/lib/sulin/dsl/d-term
 	install terminal/debian-terminal $(DESTDIR)/usr/bin/debian-terminal
 	install terminal/debian.desktop $(DESTDIR)/usr/share/applications/
